+++ src/fork/fork.ts
@@ -1,9 +1,9 @@
-import { Feed } from "../feed/feed";
+import { ObservableSource } from "../observable/observable";
 import { Query } from "../query/query";
 import { FactReference } from "../storage";
 import { Channel } from "./channel";
 
-export interface Fork extends Feed {
+export interface Fork extends ObservableSource {
     addChannel(fact: FactReference, query: Query): Channel;
     removeChannel(channel: Channel): void;
 }
