+++ src/index.ts
@@ -6,8 +6,8 @@ export { AuthorizationRules } from "./authorization/authorizationRules";
 export { Cache } from './cache';
 export { canonicalizeFact, canonicalPredecessors, computeHash } from './fact/hash';
 export { dehydrateFact, dehydrateReference, hydrate, hydrateFromTree } from "./fact/hydrate";
-export { Feed, Observable } from './feed/feed';
-export { FeedImpl } from './feed/feed-impl';
+export { ObservableSource as Feed, Observable } from './observable/observable';
+export { ObservableSourceImpl as FeedImpl } from './observable/observable-source-impl';
 export { Channel } from "./fork/channel";
 export { Fork } from "./fork/fork";
 export { PassThroughFork } from "./fork/pass-through-fork";
