+++ src/jinaga-test.ts
@@ -2,8 +2,8 @@ import { Authentication } from './authentication/authentication';
 import { AuthenticationTest } from './authentication/authentication-test';
 import { AuthorizationRules } from './authorization/authorizationRules';
 import { dehydrateFact, Dehydration } from './fact/hydrate';
-import { Feed } from './feed/feed';
-import { FeedImpl } from './feed/feed-impl';
+import { ObservableSource } from './observable/observable';
+import { ObservableSourceImpl } from './observable/observable-source-impl';
 import { SyncStatusNotifier } from './http/web-client';
 import { Jinaga } from './jinaga';
 import { MemoryStore } from './memory/memory-store';
@@ -20,7 +20,7 @@ export class JinagaTest {
   static create(config: JinagaTestConfig) {
     const store = new MemoryStore();
     this.saveInitialState(config, store);
-    const feed = new FeedImpl(store);
+    const feed = new ObservableSourceImpl(store);
     const syncStatusNotifier = new SyncStatusNotifier();
     const authentication = this.createAuthentication(config, feed);
     return new Jinaga(authentication, null, syncStatusNotifier);
@@ -37,7 +37,7 @@ export class JinagaTest {
     }
   }
 
-  static createAuthentication(config: JinagaTestConfig, inner: Feed): Authentication {
+  static createAuthentication(config: JinagaTestConfig, inner: ObservableSource): Authentication {
     const authorizationRules = config.authorization &&
       config.authorization(new AuthorizationRules());
     const userFact = config.user && dehydrateFact(config.user)[0];
